split common.rs
